+++ src/main/java/net/mcreator/spasm/block/SoakedDirtBlock.java
@@ -1,0 +1,29 @@
+package net.mcreator.spasm.block;
+
+import net.minecraft.world.level.material.MapColor;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.block.SoundType;
+import net.minecraft.world.level.block.FallingBlock;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.core.BlockPos;
+
+import com.mojang.serialization.MapCodec;
+
+public class SoakedDirtBlock extends FallingBlock {
+	public static final MapCodec<SoakedDirtBlock> CODEC = simpleCodec(SoakedDirtBlock::new);
+
+	@Override
+	public MapCodec<SoakedDirtBlock> codec() {
+		return CODEC;
+	}
+
+	@Override
+	public int getDustColor(BlockState blockstate, BlockGetter world, BlockPos pos) {
+		return blockstate.getMapColor(world, pos).col;
+	}
+
+	public SoakedDirtBlock(BlockBehaviour.Properties properties) {
+		super(properties.mapColor(MapColor.DIRT).sound(SoundType.GRAVEL).strength(0.75f, 0.5f));
+	}
+}