+++ src/main/java/net/mcreator/spasm/init/SpasmModBlocks.java
@@ -10,6 +10,7 @@ import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.Block;
 
 import net.mcreator.spasm.block.StrippedDirtBlock;
+import net.mcreator.spasm.block.SoakedDirtBlock;
 import net.mcreator.spasm.block.DirtContainerBlock;
 import net.mcreator.spasm.SpasmMod;
 
@@ -19,9 +20,11 @@ public class SpasmModBlocks {
 	public static final DeferredRegister.Blocks REGISTRY = DeferredRegister.createBlocks(SpasmMod.MODID);
 	public static final DeferredBlock<Block> STRIPPED_DIRT;
 	public static final DeferredBlock<Block> DIRT_CONTAINER;
+	public static final DeferredBlock<Block> SOAKED_DIRT;
 	static {
 		STRIPPED_DIRT = register("stripped_dirt", StrippedDirtBlock::new);
 		DIRT_CONTAINER = register("dirt_container", DirtContainerBlock::new);
+		SOAKED_DIRT = register("soaked_dirt", SoakedDirtBlock::new);
 	}
 
 	// Start of user code block custom blocks