+++ src/main/java/net/mcreator/spasm/init/SpasmModTabs.java
@@ -29,6 +29,7 @@ public class SpasmModTabs {
 			tabData.accept(SpasmModItems.WOODEN_DIRT_SHOVEL.get());
 		} else if (tabData.getTabKey() == CreativeModeTabs.BUILDING_BLOCKS) {
 			tabData.accept(SpasmModBlocks.STRIPPED_DIRT.get().asItem());
+			tabData.accept(SpasmModBlocks.SOAKED_DIRT.get().asItem());
 		} else if (tabData.getTabKey() == CreativeModeTabs.FUNCTIONAL_BLOCKS) {
 			tabData.accept(SpasmModBlocks.DIRT_CONTAINER.get().asItem());
 		}