+++ src/main/java/net/mcreator/spasm/procedures/StrippedDirtOnTickUpdateProcedure.java
@@ -1,0 +1,61 @@
+package net.mcreator.spasm.procedures;
+
+import net.minecraft.world.level.storage.TagValueInput;
+import net.minecraft.world.level.block.state.properties.Property;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.Level;
+import net.minecraft.util.ProblemReporter;
+import net.minecraft.sounds.SoundSource;
+import net.minecraft.resources.Identifier;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.core.BlockPos;
+
+import net.mcreator.spasm.init.SpasmModBlocks;
+
+public class StrippedDirtOnTickUpdateProcedure {
+	public static void execute(LevelAccessor world, double x, double y, double z) {
+		if (world.canSeeSkyFromBelowWater(BlockPos.containing(x, y, z)) && world instanceof Level _lvl1 && _lvl1.isRaining()) {
+			if (Math.random() < 0.05) {
+				if (world instanceof Level _level) {
+					if (!_level.isClientSide()) {
+						_level.playSound(null, BlockPos.containing(x, y, z), BuiltInRegistries.SOUND_EVENT.getValue(Identifier.parse("spasm:block.stripped_dirt.soaked")), SoundSource.BLOCKS, 1, 1);
+					} else {
+						_level.playLocalSound(x, y, z, BuiltInRegistries.SOUND_EVENT.getValue(Identifier.parse("spasm:block.stripped_dirt.soaked")), SoundSource.BLOCKS, 1, 1, false);
+					}
+				}
+				{
+					BlockPos _bp = BlockPos.containing(x, y, z);
+					BlockState _bs = SpasmModBlocks.SOAKED_DIRT.get().defaultBlockState();
+					BlockState _bso = world.getBlockState(_bp);
+					for (Property<?> _propertyOld : _bso.getProperties()) {
+						Property _propertyNew = _bs.getBlock().getStateDefinition().getProperty(_propertyOld.getName());
+						if (_propertyNew != null && _bs.getValue(_propertyNew) != null)
+							try {
+								_bs = _bs.setValue(_propertyNew, _bso.getValue(_propertyOld));
+							} catch (Exception e) {
+							}
+					}
+					BlockEntity _be = world.getBlockEntity(_bp);
+					CompoundTag _bnbt = null;
+					if (_be != null) {
+						_bnbt = _be.saveWithFullMetadata(world.registryAccess());
+						_be.setRemoved();
+					}
+					world.setBlock(_bp, _bs, 3);
+					if (_bnbt != null) {
+						_be = world.getBlockEntity(_bp);
+						if (_be != null) {
+							try {
+								_be.loadWithComponents(TagValueInput.create(ProblemReporter.DISCARDING, world.registryAccess(), _bnbt));
+							} catch (Exception ignored) {
+							}
+						}
+					}
+				}
+			}
+		}
+	}
+}